Refactor ToggableGroup and Prices components for improved UI consistency and flexibility
Adjust ToggableGroup.vue to add a new prop 'small' that conditionally
applies the '-small' class to the label element. This change allows for
more flexibility in styling the component and ensures UI consistency
with other components.

In Prices.vue, replace the background image source for crypto logos with
the 'logo' property from the 'market' object. This modification
simplifies the code and makes it easier to manage and update crypto
logos in the future.

In PricesSettings.vue, update the checkbox labels to include the
'-small' class for consistent styling with other checkboxes in the
component. Additionally, modify the 'toggable-section__title' template
to display the selected volume threshold and period values as badges,
providing at-a-glance information to users.

These changes enhance the visual consistency and maintainability of the
components and improve the overall user experience.